+++ AgreementMaker-OSGi/AgreementMaker-Core/pom.xml
@@ -175,6 +175,12 @@
 			<version>0.0.1-SNAPSHOT</version>
 			<scope>provided</scope>
 		</dependency>
+
+        <dependency>
+            <groupId>edu.uic.cs.advis.am</groupId>
+            <artifactId>AgreementMaker-SimilarityMetrics</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+        </dependency>
 	
 		<dependency>
 			<groupId>commons-lang</groupId>
@@ -266,12 +272,6 @@
 			<groupId>simpack</groupId>
 			<artifactId>simpack</artifactId>
 			<version>0.91</version>
-		</dependency>
-
-		<dependency>
-			<groupId>jaws</groupId>
-			<artifactId>jaws</artifactId>
-			<version>1.3</version>
 		</dependency>
 
 		<dependency>
