+++ AgreementMaker-OSGi/AgreementMaker-Matchers/src/main/java/am/matcher/LexicalSynonymMatcher/LexicalSynonymMatcher.java
@@ -9,7 +9,6 @@ import org.apache.log4j.Logger;
 import am.app.Core;
 import am.app.lexicon.Lexicon;
 import am.app.lexicon.LexiconSynSet;
-import am.app.lexicon.subconcept.SynonymTermLexicon;
 import am.app.mappingEngine.AbstractMatcher;
 import am.app.mappingEngine.AbstractMatcherParametersPanel;
 import am.app.mappingEngine.LexiconStore.LexiconRegistry;
@@ -18,7 +17,7 @@ import am.app.mappingEngine.MatcherFeature;
 import am.app.mappingEngine.similarityMatrix.ArraySimilarityMatrix;
 import am.app.mappingEngine.similarityMatrix.SimilarityMatrix;
 import am.app.ontology.Node;
-import am.utility.WordNetUtils;
+import am.app.wordnet.WordNetUtils;
 
 import com.hp.hpl.jena.ontology.OntResource;
