+++ AgreementMaker-OSGi/AgreementMaker-SimilarityMetrics/pom.xml
@@ -66,9 +66,8 @@
 	<dependencies>
         <dependency>
             <groupId>edu.uic.cs.advis.am</groupId>
-            <artifactId>AgreementMaker-Core</artifactId>
-            <version>0.3.0-SNAPSHOT</version>
-            <scope>provided</scope>
+            <artifactId>AgreementMaker-WordNet</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
         </dependency>
 		<dependency>
 			<groupId>simmetrics</groupId>
