+++ AgreementMaker-OSGi/pom.xml
@@ -44,6 +44,7 @@
         <module>Matcher-BaseSimilarity</module>
         <module>Matcher-AdvancedSimilarity</module>
         <module>AgreementMaker-Matchers</module>
+        <module>AgreementMaker-WordNet</module>
         <module>Matcher-PRA</module>
         <module>Matcher-OAEI</module>
         <module>Matcher-LinkedOpenData</module>
